commit 2

+++ schedule/src/main/java/Services/ClientService.java
@@ -5,10 +5,10 @@ import model.Client;
 import java.util.List;
 
 public interface ClientService {
-    Client findById(int id);
+    Client findById(Long id);
     Client findByName(String username);
     Client add(Client client);
     List<Client> listAll();
-    void remove(int id);
+    void remove(Long id);
     Client update(Client client);
 }
